[level 3] Title: 입국심사, Time: 116.82 ms, Memory: 109 MB -BaekjoonHub

+++ 프로그래머스/3/43238. 입국심사/입국심사.java
@@ -17,10 +17,10 @@ class Solution {
         return min;
     }
     public boolean isAble(long mid, int n, int[] times) {
-        long count = 0;
+        int count = 0;
         for(int i = 0; i< times.length;i++){
             int t = times[i];
-            count += (long)Math.floor((long)mid / t);
+            count += (int)Math.floor((long)mid / t);
             if(count >= n){
                 return true;
             }
